pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "mkchimeenv"
authors = [{name = "Richard Shaw", email = "richard@phas.ubc.ca"}]
license = {file = "LICENSE"}
readme = "README.md"
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python = ">=3.8"
dependencies = [
    "requirements-detector",
    "virtualenv",
    "virtualenv-api",
    "GitPython",
    "rich",
    "click >= 8.0",
]

[project.scripts]
mkchimeenv = "mkchimeenv:cli"
